Validators: corrigindo erros e adicionando Validator de user para update Nomes das classes sendo concertadas routes.js atualizado
