+++ src/routes.js
@@ -26,7 +26,7 @@ import AuthMiddleware from './app/middlewares/auth';
 import RolesMiddleware from './app/middlewares/roles';
 
 // validators
-import CreateUserValidator from './app/middlewares/validators/CreateUser';
+import UserValidator from './app/middlewares/validators/User';
 import ManifestationValidator from './app/middlewares/validators/Manifestation';
 import UserLoginValidator from './app/middlewares/validators/UserLogin';
 import GenericValidator from './app/middlewares/validators/Generic';
@@ -57,7 +57,7 @@ router.get('/user/:id?*', UserController.fetch);
  * somente para criação de usuários admin
  */
 
-router.post('/user', CreateUserValidator, UserController.save);
+router.post('/user', UserValidator.save, UserController.save);
 router.post('/auth', UserLoginValidator, AuthController.login);
 
 /**
